Refactor widgets folder, add new assets
