--- app/src/main/java/widget/large/MissionWidgetLargeReceiver.kt
@@ -1,8 +1,0 @@
-package widget.large
-
-import androidx.glance.appwidget.GlanceAppWidget
-import widget.MissionWidgetReceiver
-
-class MissionWidgetLargeReceiver : MissionWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = MissionWidgetLarge()
-}
