--- app/src/main/java/widget/minimal/MissionWidgetMinimalReceiver.kt
@@ -1,8 +1,0 @@
-package widget.minimal
-
-import androidx.glance.appwidget.GlanceAppWidget
-import widget.MissionWidgetReceiver
-
-class MissionWidgetMinimalReceiver : MissionWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = MissionWidgetMinimal()
-}
