+++ app/src/main/java/widget/missions/large/MissionWidgetLargeReceiver.kt
@@ -1,0 +1,8 @@
+package widget.large
+
+import androidx.glance.appwidget.GlanceAppWidget
+import widget.MissionWidgetReceiver
+
+class MissionWidgetLargeReceiver : MissionWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = MissionWidgetLarge()
+}
