+++ app/src/main/java/widget/missions/minimal/MissionWidgetMinimalReceiver.kt
@@ -1,0 +1,8 @@
+package widget.minimal
+
+import androidx.glance.appwidget.GlanceAppWidget
+import widget.MissionWidgetReceiver
+
+class MissionWidgetMinimalReceiver : MissionWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = MissionWidgetMinimal()
+}
